refactor(datasource): rename datasource_http
Rename datasource_http module to datasource_ncei.
