+++ src/goesdl/datasource/datasource_ncei.py
